refactor(canvas): extract canvasGeometry — break GRID_SIZE boundary leak
/lowy audit of `packages/client/src/canvas/` flagged one concrete
violation: tile-packing modules (`repoIslands.ts`, `tilePlacement.ts`)
import `GRID_SIZE` and `snapToGrid` from `viewport/transforms.ts`.
Those values are part of the viewport's internal coordinate
system; the packing algorithm should not depend on viewport
internals.

Today it works because tile-space and viewport-space share the
same grid. If they ever needed to diverge, this would be a silent
conflict cut through `repoIslands.ts` as a surprise. More
immediately, it blocks the extraction `/lowy` identified —
`repoIslands.ts` + `tilePlacement.ts` are the first cleanly
framework-shaped candidate in the canvas cluster (a future
`@kolu/canvas-layout` for "2D canvas with grouped tiles"). The
viewport import is the only thing keeping them from being
domain-free.

Fix: `canvas/canvasGeometry.ts` defines `GRID_SIZE` + `snapToGrid`.
`viewport/transforms.ts` re-exports them so the gesture/animatedPan
viewport-internal callers keep their local-feeling import.
Packing modules now import from the neutral location.

No behavior change. 172/172 client unit tests pass.
