+++ docs/architecture-ralph-report.md
@@ -104,7 +104,8 @@ The repo has clearly internalized the extraction discipline:
 |     2 | `Terminal.tsx` (930 LOC, mixed concerns) | extract `mobileTouch.ts` — `setupMobileTapToFocus(term)` + `setupMobileTouchScroll(container, getTerm)` | 0 lint Δ; isolates iOS-Safari platform volatility (Lowy Axis 8) | Terminal.tsx 930→821 (−109); `mobileTouch.ts` +135 | `0893f3d2` | /lowy audit said the proposed `@kolu/solid-xterm` framework extraction *fails* the reuse test (one consumer; would shape API around implementation, not stable contract). What it *did* recommend was the mobile-touch state machine: 110 LOC of iOS-tap-vs-scroll heuristic + xterm-touchscroll bridge with a real platform-volatility axis. The receptacle is a single file; the axis (iOS Safari focus/touch behavior) now has a documented home for the next workaround. 172/172 client unit tests pass. |
 |     3 | xterm WebGL lifecycle inside `Terminal.tsx` | create **new package** `@kolu/solid-xterm`; move `loadWebgl`/`unloadWebgl`/`clearTextureAtlas`/`webgl?.textureAtlas` probe into `createXtermWebgl(getTerm, hooks)` | 0 lint Δ; isolates xterm WebGL addon + Chrome GPU context volatility; establishes `@kolu/solid-xterm` shape | Terminal.tsx 821→779 (−42); new package +205 (`src/webgl.ts` 154 + `src/index.ts` 19 + README 32) | `df3f152f` | **Reversal of the cycle-2 /lowy verdict.** User feedback: "single consumer is not a good excuse — surface has a single consumer too. just like electricity even if it used in only one home". Re-extracted under the Surface precedent: encapsulate a stable axis (xterm WebglAddon API + Chrome's ~16 per-tab GPU context budget + link-layer-canvas selector trap + `WEBGL_lose_context.loseContext()` ordering) behind a small factory the host plugs hooks into. `webglTracker.ts` stays in client/ (Kolu-specific #591 debug ledger); the framework calls `onCreate`/`onLoseContextCalled`/`onDispose` hooks the host wires into it. First step toward a multi-cycle `@kolu/solid-xterm` build-out. 172/172 client unit tests pass; pnpm hash unchanged. |
 |     4 | reactive xterm option sync (theme + fontSize) in `Terminal.tsx` | grow `@kolu/solid-xterm`: add `attachXtermStyleSync(getTerm, { theme, fontSize, onThemeChange, onFontSizeChange })` export + `./style-sync` entry point | 0 lint Δ; second framework axis encapsulated (xterm options live-write + after-change hook ordering) | Terminal.tsx 779→770 (−9); framework +73 (`src/styleSync.ts`) | `510b1e54` | Two `createEffect(on(..., { defer: true }))` blocks + their atlas-clear / refit follow-ups now sit behind one helper. The `defer: true` discipline (initial values come from XTerm constructor, only subsequent reactive changes flow through) lives in the framework instead of being re-discovered per call site. Separate `onThemeChange` / `onFontSizeChange` hooks because the theme axis has no fit implication — collapsing to one `afterChange` would force unnecessary work on theme swap. 172/172 client unit tests pass. |
-|     5 | `client/src/scrollLock.ts` (zero-Kolu-coupling 120-LOC primitive) | move it bit-for-bit into `@kolu/solid-xterm` as `./scroll-lock`; drop the client-side copy | 0 lint Δ; consolidates third xterm-volatility axis (scrollback freeze-write + buffer.active.baseY/viewportY math) in the framework | Terminal.tsx import path change; net file count: −1 in client, +1 in solid-xterm | (cycle 5) | The file was already at the right altitude — `createScrollLock(enabled)` only depended on `@xterm/xterm` and `solid-js`. The new framework is its natural home. Reduces the count of "things in client/src/ that should be elsewhere" by one and grows `@kolu/solid-xterm`'s API surface from 2 → 3 exports. 172/172 client unit tests pass. |
+|     5 | `client/src/scrollLock.ts` (zero-Kolu-coupling 120-LOC primitive) | move it bit-for-bit into `@kolu/solid-xterm` as `./scroll-lock`; drop the client-side copy | 0 lint Δ; consolidates third xterm-volatility axis (scrollback freeze-write + buffer.active.baseY/viewportY math) in the framework | Terminal.tsx import path change; net file count: −1 in client, +1 in solid-xterm | `fa8ce514` | The file was already at the right altitude — `createScrollLock(enabled)` only depended on `@xterm/xterm` and `solid-js`. The new framework is its natural home. Reduces the count of "things in client/src/ that should be elsewhere" by one and grows `@kolu/solid-xterm`'s API surface from 2 → 3 exports. 172/172 client unit tests pass. |
+|     6 | canvas: `GRID_SIZE` + `snapToGrid` leaking from `viewport/transforms.ts` to `repoIslands.ts` + `tilePlacement.ts` | extract `canvasGeometry.ts`; viewport re-exports for backward compat; packing imports from neutral location | breaks one Lowy boundary violation; unblocks future `@kolu/canvas-layout` extraction identified by /lowy | +1 file (`canvasGeometry.ts`, 25 LOC); net repo LOC ≈ +5 | (cycle 6) | /lowy audit (canvas cluster, ≈800-word report) flagged one violation: tile-packing modules (`repoIslands.ts`, `tilePlacement.ts`) reach into `viewport/transforms.ts` for `GRID_SIZE`/`snapToGrid`. The viewport's internal coordinate constants shouldn't be tile-packing's dependency surface — they're sharing today but if tile-space ever needs a different grid the conflict is silent. The audit also identified `repoIslands.ts` + `tilePlacement.ts` as **the first extractable framework candidate from the canvas cluster** (a future `@kolu/canvas-layout` for "2D canvas with grouped tiles"), blocked precisely by this import. Fix the boundary now; extraction can follow without re-touching it. 172/172 client unit tests pass. |
 
 ---
