+++ packages/client/src/canvas/viewport/transforms.ts
@@ -1,20 +1,22 @@
 /** Pure viewport math — takes values, returns values. No signals, no DOM.
  *  Encapsulates the zoom/pan algorithm so it can evolve (easing, constraints,
- *  undo) without touching gesture input or CSS generation. */
+ *  undo) without touching gesture input or CSS generation.
+ *
+ *  `GRID_SIZE` / `snapToGrid` live in `../canvasGeometry.ts` — shared
+ *  between viewport math and tile-packing so the packing algorithms
+ *  don't have to reach inside the viewport boundary. Re-exported here
+ *  for the viewport-internal callers (gestures, animatedPan,
+ *  useCanvasViewport) so they keep their local-feeling import. */
+
+export { GRID_SIZE, snapToGrid } from "../canvasGeometry";
 
 export const MIN_ZOOM = 0.15;
 export const MAX_ZOOM = 3;
-export const GRID_SIZE = 24;
 const ZOOM_STEP = 1.25;
 
 /** Clamp zoom to allowed range. */
 export function clampZoom(z: number): number {
   return Math.min(MAX_ZOOM, Math.max(MIN_ZOOM, z));
-}
-
-/** Snap a value to the canvas grid. */
-export function snapToGrid(value: number): number {
-  return Math.round(value / GRID_SIZE) * GRID_SIZE;
 }
 
 /** Divide a screen-space delta by zoom for canvas-space positioning. */
